Replace _MessageSearchProvider with MessageSearchProvider
