+++ src/modules/MessageSearch/context/MessageSearchProvider.tsx
@@ -1,26 +1,19 @@
-import React, {
-  useRef,
-  useState,
-  useReducer,
-} from 'react';
-import { SendbirdError } from '@sendbird/chat';
-import type { MessageSearchQuery } from '@sendbird/chat/message';
+import React, { createContext, useRef, useContext, useCallback, useEffect } from 'react';
 import type { GroupChannel } from '@sendbird/chat/groupChannel';
+import { MessageSearchQuery } from '@sendbird/chat/message';
+import { ClientSentMessages } from '../../../types';
+import { SendbirdError } from '@sendbird/chat';
 import type { MessageSearchQueryParams } from '@sendbird/chat/lib/__definition';
 
 import useSendbirdStateContext from '../../../hooks/useSendbirdStateContext';
-import { ClientSentMessages } from '../../../types';
-
-import messageSearchReducer from './dux/reducers';
-import messageSearchInitialState, { State as MessageSearchReducerState } from './dux/initialState';
 
 import useSetChannel from './hooks/useSetChannel';
 import useGetSearchMessages from './hooks/useGetSearchedMessages';
-import useScrollCallback, {
-  CallbackReturn as UseScrollCallbackType,
-} from './hooks/useScrollCallback';
+import useScrollCallback from './hooks/useScrollCallback';
 import useSearchStringEffect from './hooks/useSearchStringEffect';
 import { CoreMessageType } from '../../../utils';
+import { createStore } from '../../../utils/storeManager';
+import { useStore } from '../../../hooks/useStore';
 
 export interface MessageSearchProviderProps {
   channelUrl: string;
@@ -31,83 +24,72 @@ export interface MessageSearchProviderProps {
   onResultClick?(message: ClientSentMessages): void;
 }
 
-interface MessageSearchProviderInterface extends MessageSearchProviderProps {
-  requestString?: string;
-  retryCount: number;
-  setRetryCount: React.Dispatch<React.SetStateAction<number>>;
-  selectedMessageId: number;
-  setSelectedMessageId: React.Dispatch<React.SetStateAction<number>>;
-  messageSearchDispatcher: (props: { type: string, payload: any }) => void;
-  scrollRef: React.RefObject<HTMLDivElement>;
-  allMessages: MessageSearchReducerState['allMessages'];
+export interface MessageSearchState extends MessageSearchProviderProps {
+  channelUrl: string;
+  allMessages: ClientSentMessages[];
   loading: boolean;
   isInvalid: boolean;
-  currentChannel: GroupChannel;
-  currentMessageSearchQuery: MessageSearchQuery;
+  initialized: boolean;
+  currentChannel: GroupChannel | null;
+  currentMessageSearchQuery: MessageSearchQuery | null;
   hasMoreResult: boolean;
-  onScroll: UseScrollCallbackType;
-  handleRetryToConnect: () => void;
-  handleOnScroll: (e: React.BaseSyntheticEvent) => void;
+  retryCount: number;
+  selectedMessageId: number | null;
+  requestString: string;
+  onScroll?: ReturnType<typeof useScrollCallback>;
+  handleOnScroll?: (e: React.BaseSyntheticEvent) => void;
+  scrollRef?: React.RefObject<HTMLDivElement>;
 }
 
-const MessageSearchContext = React.createContext<MessageSearchProviderInterface | null>(null);
+const initialState: MessageSearchState = {
+  channelUrl: '',
+  allMessages: [],
+  loading: false,
+  isInvalid: false,
+  initialized: false,
+  currentChannel: null,
+  currentMessageSearchQuery: null,
+  messageSearchQuery: null,
+  hasMoreResult: false,
+  retryCount: 0,
+  selectedMessageId: null,
+  searchString: '',
+  requestString: '',
+};
 
-const MessageSearchProvider: React.FC<MessageSearchProviderProps> = (props: MessageSearchProviderProps) => {
-  const {
-    // message search props
-    channelUrl,
-    searchString,
-    messageSearchQuery,
-    onResultLoaded,
-    onResultClick,
-  } = props;
+export const MessageSearchContext = createContext<ReturnType<typeof createStore<MessageSearchState>> | null>(null);
 
-  const globalState = useSendbirdStateContext();
-
-  // hook variables
-  const [retryCount, setRetryCount] = useState(0); // this is a trigger flag for activating useGetSearchMessages
-  const [selectedMessageId, setSelectedMessageId] = useState(0);
-  const [messageSearchStore, messageSearchDispatcher] = useReducer(messageSearchReducer, messageSearchInitialState);
-  const {
-    allMessages,
-    loading,
-    isInvalid,
-    currentChannel,
-    currentMessageSearchQuery,
-    hasMoreResult,
-  } = messageSearchStore;
-
-  const logger = globalState?.config?.logger;
-  const sdk = globalState?.stores?.sdkStore?.sdk;
-  const sdkInit = globalState?.stores?.sdkStore?.initialized;
+const MessageSearchManager: React.FC<MessageSearchProviderProps> = ({
+  channelUrl,
+  searchString,
+  messageSearchQuery,
+  onResultLoaded,
+  onResultClick,
+}) => {
+  const { state, updateState } = useMessageSearchStore();
+  const { config, stores } = useSendbirdStateContext();
+  const sdk = stores?.sdkStore?.sdk;
+  const sdkInit = stores?.sdkStore?.initialized;
+  const { logger } = config;
   const scrollRef = useRef<HTMLDivElement>(null);
-  const handleOnScroll = (e: React.BaseSyntheticEvent) => {
-    const scrollElement = e.target as HTMLDivElement;
-    const {
-      scrollTop,
-      scrollHeight,
-      clientHeight,
-    } = scrollElement;
-
-    if (!hasMoreResult) {
-      return;
-    }
-    if (scrollTop + clientHeight >= scrollHeight) {
-      onScroll(() => {
-        // after load more searched messages
-      });
-    }
-  };
 
   useSetChannel(
     { channelUrl, sdkInit },
     { sdk, logger },
   );
 
-  const requestString = useSearchStringEffect({ searchString: searchString ?? '' });
+  const requestString = useSearchStringEffect(
+    { searchString: searchString ?? '' },
+  );
 
   useGetSearchMessages(
-    { currentChannel, channelUrl, requestString, messageSearchQuery, onResultLoaded },
+    {
+      currentChannel: state.currentChannel,
+      channelUrl,
+      requestString,
+      messageSearchQuery,
+      onResultLoaded,
+    },
     { sdk, logger },
   );
 
@@ -116,40 +98,72 @@ const MessageSearchProvider: React.FC<MessageSearchProviderProps> = (props: Mess
     { logger },
   );
 
-  const handleRetryToConnect = () => {
-    setRetryCount(retryCount + 1);
-  };
-  return (
-    <MessageSearchContext.Provider value={{
+  const handleOnScroll = useCallback((e: React.BaseSyntheticEvent) => {
+    const scrollElement = e.target as HTMLDivElement;
+    const { scrollTop, scrollHeight, clientHeight } = scrollElement;
+
+    if (!state.hasMoreResult) {
+      return;
+    }
+    if (scrollTop + clientHeight >= scrollHeight) {
+      onScroll(() => {
+        // after load more searched messages
+      });
+    }
+  }, [state.hasMoreResult, onScroll]);
+
+  useEffect(() => {
+    updateState({
       channelUrl,
       searchString,
-      requestString,
       messageSearchQuery,
-      onResultLoaded,
       onResultClick,
-      retryCount,
-      setRetryCount,
-      selectedMessageId,
-      setSelectedMessageId,
-      messageSearchDispatcher,
-      allMessages,
-      loading,
-      isInvalid,
-      currentChannel,
-      currentMessageSearchQuery,
-      hasMoreResult,
       onScroll,
-      scrollRef,
-      handleRetryToConnect,
       handleOnScroll,
-    }}>
-      {props?.children}
+      scrollRef,
+      requestString,
+    });
+  }, [channelUrl, searchString, messageSearchQuery, onResultClick, updateState, requestString]);
+
+  return null;
+};
+
+const createMessageSearchStore = () => createStore(initialState);
+const InternalMessageSearchProvider: React.FC<React.PropsWithChildren<unknown>> = ({ children }) => {
+  const storeRef = useRef(createMessageSearchStore());
+
+  return (
+    <MessageSearchContext.Provider value={storeRef.current}>
+      {children}
     </MessageSearchContext.Provider>
   );
 };
 
+const MessageSearchProvider: React.FC<MessageSearchProviderProps> = ({
+  children,
+  channelUrl,
+  searchString,
+  messageSearchQuery,
+  onResultLoaded,
+  onResultClick,
+}) => {
+
+  return (
+    <InternalMessageSearchProvider>
+      <MessageSearchManager
+        channelUrl={channelUrl}
+        searchString={searchString}
+        messageSearchQuery={messageSearchQuery}
+        onResultLoaded={onResultLoaded}
+        onResultClick={onResultClick}
+      />
+      {children}
+    </InternalMessageSearchProvider>
+  );
+};
+
 const useMessageSearchContext = () => {
-  const context = React.useContext(MessageSearchContext);
+  const context = useContext(MessageSearchContext);
   if (!context) throw new Error('MessageSearchContext not found. Use within the MessageSearch module.');
   return context;
 };
@@ -157,4 +171,13 @@ const useMessageSearchContext = () => {
 export {
   MessageSearchProvider,
   useMessageSearchContext,
+  MessageSearchManager,
+};
+
+/**
+ * A specialized hook for MessageSearch state management
+ * @returns {ReturnType<typeof createStore<MessageSearchState>>}
+ */
+const useMessageSearchStore = () => {
+  return useStore(MessageSearchContext, state => state, initialState);
 };
